// src/api/endpoints/model/inspectionSpec.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * (title)
 * OpenAPI spec version: 0.0.0
 */

export interface InspectionSpec {
  code: string;
  tenantId: string;
  inspectionObjectCode?: string;
  name: string;
  remark?: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}