// src/api/endpoints/model/inspectionObject.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * (title)
 * OpenAPI spec version: 0.0.0
 */

export interface InspectionObject {
  code: string;
  inspectionSpecialtyCode: string;
  sourceProjectNo: string;
  sourceProjectName: string;
  name: string;
  isOptionalForQualification: boolean;
  isOfficial: boolean;
  enabled: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}